+++ main.c
@@ -9,9 +9,9 @@ int main(int argc, char *argv[]) {
     int order;
     char name[25];
     char eq[1];
-    float epsilon;
+    double epsilon;
 
-    fscanf(stdin, "%s%s%e", name,eq, &epsilon);
+    fscanf(stdin, "%s%s%lf", name,eq, &epsilon);
 
     fscanf(stdin, "%s%s%i", name,eq, &order);
 
@@ -20,13 +20,13 @@ int main(int argc, char *argv[]) {
 
     for(int i=0;i<=order;i++) {
         complex_num* cn =(complex_num*) malloc(sizeof(complex_num));
-        fscanf(stdin, "%s%i%s%e%e", name,&indx,eq, &cn->real,&cn->imaginary);
+        fscanf(stdin, "%s%i%s%lf%lf", name,&indx,eq, &cn->real,&cn->imaginary);
         coeff[indx] = cn;
     }
     complex_num** deriv = deriv_coeff(coeff,order);
 
     complex_num* initial =(complex_num*) malloc(sizeof(complex_num));
-    fscanf(stdin, "%s%s%e%e", name,eq, &initial->real,&initial->imaginary);
+    fscanf(stdin, "%s%s%lf%lf", name,eq, &initial->real,&initial->imaginary);
 
     printf("inputs:\n");
     print_input(order,epsilon,coeff,initial);
@@ -56,7 +56,7 @@ int main(int argc, char *argv[]) {
 
 void print_input(int order, double epsilon, complex_num** coeff, complex_num* initial){
     printf("order: %i\n",order);
-    printf("epsilon: %e\n",epsilon);
+    printf("epsilon: %.16e\n",epsilon);
     for(int i=0;i<=order;i++){
         printf("coeff[%i]: ",i);
         print_complex(coeff[i]);
@@ -68,7 +68,7 @@ void print_input(int order, double epsilon, complex_num** coeff, complex_num* in
 }
 
 void print_complex(complex_num* num){
-    printf("%e %e",num->real,num->imaginary);
+    printf("%.16e %.16e",num->real,num->imaginary);
 }
 
 complex_num* pow_copmplex( complex_num* num1, int power){
